Move src/Listener to src/EventListener. Add permission for route home
