+++ config/permissions.yaml
@@ -189,6 +189,8 @@
     - 100
 "holiday_info.delete":
     - 100
+"home":
+    - 100
 "index":
     - 100
 "legalnotices":
