+++ src/EventListener/ControllerAuthorizationListener.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Listener;
+namespace App\EventListener;
 
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Event\GetResponseEvent;
